+++ src/core/custom-button/custom-button.styled.tsx
@@ -2,10 +2,12 @@ import * as R from 'ramda'
 import styled from 'styled-components/native'
 import { COLOR } from '../colors/colors.constants'
 import { TEXT_SIZE } from '../text/text.constants'
+import { appendStyleWhenProvided } from '@core/styled';
 
 interface CustomButtonContainerProps {
   invisible?: boolean;
   color?: COLOR;
+  minHeight?: number;
 }
 
 const getCustomButtonContainerOpacity = R.pipe(
@@ -18,7 +20,7 @@ const getCustomButtonContainerBackgroundColor = R.propOr(COLOR.DODGER_BLUE, 'col
 
 export const CustomButtonContainer = styled.TouchableOpacity.attrs<CustomButtonContainerProps>(({ invisible }) => ({
   activeOpacity: invisible ? 0 : undefined,
-}))`
+}))<CustomButtonContainerProps>`
   justify-content: center;
   align-items: center;
   align-self: center;
@@ -26,6 +28,7 @@ export const CustomButtonContainer = styled.TouchableOpacity.attrs<CustomButtonC
   border-radius: 30px;
   background-color: ${getCustomButtonContainerBackgroundColor};
   opacity: ${getCustomButtonContainerOpacity};
+  ${appendStyleWhenProvided('min-height', 'minHeight', 'px')};
 `
 
 const getCustomButtonTitleFontSize = R.propOr(TEXT_SIZE.M, 'titleSize')