+++ src/core/header/header.tsx
@@ -1,12 +1,13 @@
 import * as React from 'react'
-import { StatusBar, TouchableOpacity } from 'react-native'
+import { StatusBar } from 'react-native'
 import { useSafeAreaInsets } from 'react-native-safe-area-context'
 import { useNavigation } from '@react-navigation/native'
 import { genericLightShadow, genericTextShadow } from '@core/shadow/shadow.constants'
-import { COLOR } from '@core/colors/colors.constants'
 import { useLocalize } from '@core/hooks/use-localize.hook'
 import { isPlatform } from '@core/is-platform/is-platform'
-import { HeaderSideContentConfig, HeaderType } from './header.models'
+import { ScreenType } from '@core/models'
+import { screenTypeToColorMap } from '@core/maps/screen-type-to-color-map'
+import { HeaderSideContentConfig } from './header.models'
 import { useHeaderTextSize } from './hooks/use-header-text-size.hook'
 import {
   BackButtonContainer, BackButtonIcon, HeaderContainer, HeaderLeftButtonContainer, HeaderLeftButtonIndicator,
@@ -14,7 +15,7 @@ import {
 } from './header.styled'
 
 export interface HeaderProps {
-  type: HeaderType;
+  type: ScreenType;
   title?: string;
   backButton?: boolean;
   backButtonAlert?: Function;
@@ -29,12 +30,7 @@ export const Header = ({ type, title, backButton, backButtonAlert, onTouchEnd, r
   const localize = useLocalize()
   const { onHeaderTextLayout, headerTextSize } = useHeaderTextSize()
 
-  const color: COLOR = {
-    dashboard: COLOR.FIRE_BRICK,
-    dictionary: COLOR.DODGER_BLUE,
-    charade: COLOR.DARK_SEA_GREEN,
-    more: COLOR.GOLD,
-  }[type]
+  const color = screenTypeToColorMap[type]
 
   const onBackPress = () => {
     if (backButtonAlert) {