+++ src/core/models.ts
@@ -1,1 +1,3 @@
 export type NumberFlag = 0 | 1
+
+export type ScreenType = 'more' | 'charade' | 'dictionary' | 'dashboard'